requirements.txt
aiosqlite==0.19.0
anyio==3.7.0
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
makefun==1.15.1
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.9
PyJWT==2.6.0
pytest==7.3.2
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.3
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
